+++ hnote/Screens/MainScreen.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { View, Text } from 'react-native';
+import MainTemplate from '../components/templates/MainTemplate';
+
+const MainScreen = props => {
+    <View>
+        <Text>MainScreen</Text>
+        <MainTemplate></MainTemplate>
+    </View>;
+};
+
+export default MainScreen;